+++ src/app/swiss/swiss.component.ts
@@ -1,0 +1,6 @@
+import { Component } from '@angular/core';
+
+@Component({
+  templateUrl: './swiss.component.html'
+})
+export class SwissComponent { }