+++ src/app/swiss/swiss.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 
+import { SwissComponent } from './swiss.component';
 import { SwissPlayersComponent } from './swiss-players.component';
 
 @NgModule({
   declarations: [
+    SwissComponent,
     SwissPlayersComponent
   ]
 })